H2 の実行時にテーブルが作成されない現象を解消する

+++ src/main/java/com/kasakaid/jpaandquerydsl/domain/artist/Band.java
@@ -9,7 +9,7 @@ import javax.persistence.Entity;
 @NoArgsConstructor
 @DiscriminatorValue("band")
 public class Band extends Artist {
-    public Band(long id, String name) {
+    public Band(int id, String name) {
         super(id, name);
     }
 }
